python/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3

# AI APIs
openai==1.10.0
anthropic==0.8.1

# Utils
requests==2.31.0
aiofiles==23.2.1

langchain
langchain-community
langchain-chroma
langchain-huggingface
langchain-groq
langchain-openai
chromadb